+++ demo/config/shrine.rb
@@ -48,9 +48,9 @@ if ENV["RACK_ENV"] == "production"
     type     = request.params["type"]
 
     {
-      content_disposition:    "inline; filename=\"#{filename}\"", # set download filename
-      content_type:           type,                               # set content type (defaults to "application/octet-stream")
-      content_length_range:   0..(10*1024*1024),                  # limit upload size to 10 MB
+      content_disposition:    ContentDisposition.inline(filename), # set download filename
+      content_type:           type,                                # set content type (defaults to "application/octet-stream")
+      content_length_range:   0..(10*1024*1024),                   # limit upload size to 10 MB
     }
   }
 else