+++ lib/shrine/plugins/default_url_options.rb
@@ -5,14 +5,14 @@ class Shrine
     # The `default_url_options` plugin allows you to specify URL options that
     # will be applied by default for uploaded files of specified storages.
     #
-    #     plugin :default_url_options, store: {download: true}
+    #     plugin :default_url_options, store: { download: true }
     #
     # You can also generate the default URL options dynamically by using a
     # block, which will receive the UploadedFile object along with any options
     # that were passed to `UploadedFile#url`.
     #
     #     plugin :default_url_options, store: ->(io, **options) do
-    #       {response_content_disposition: "attachment; filename=\"#{io.original_filename}\""}
+    #       { response_content_disposition: ContentDisposition.attachment(io.original_filename) }
     #     end
     #
     # In both cases the default options are merged with options passed to